+++ CHANGES.md
@@ -11,6 +11,10 @@ Released: -<br>Codename: Gongqing
   value of `arg_name` ([issue #427][issue_427]).
 - Add `FileSchema` to generate OpenAPI file response ([issue #447][issue_447]).
 - Support using `{}` to represent not only empty body (204) but also empty schema.
+- Remove the previously deprecated code:
+  - The `tag` parameter in `@app.doc`.
+  - The `role` parameter in `@app.auth_required`.
+  - The `redoc_path` parameter in `apiflask.APIFlask` and the `/redoc` path.
 
 [issue_427]: https://github.com/apiflask/apiflask/issues/427
 [issue_442]: https://github.com/apiflask/apiflask/issues/442